+++ website/src/documentation/index.ts
@@ -1,4 +1,5 @@
 import type { Doc } from "../components/docs";
+import { Anim_Docs } from "./anim";
 import { Delay_Docs } from "./delay";
 import { Rect_Docs } from "./rect";
 import { Vec2_Docs } from "./vec2";
@@ -6,5 +7,6 @@ import { Vec2_Docs } from "./vec2";
 export const DOCS: Doc[] = [
     ...Vec2_Docs,
     ...Rect_Docs,
-    ...Delay_Docs
+    ...Delay_Docs,
+    ...Anim_Docs
 ]