+++ CHANGELOG.md
@@ -6,9 +6,10 @@ This project adheres to Semantic Versioning.
 
 ## [Unstaged]
 
-- Delay Timing API added \([delay.ts](src/delay.ts)\)
-- Rectangle API added \([rect.ts](src/math/rect.ts)\)
-- Vector 2 API added \([vec2.ts](src/math/vec2.ts)\)
+- Added AnimChannel Animation helper \([anim.ts](src/anim.ts)\)
+- Added Delay Timing helper \([delay.ts](src/delay.ts)\)
+- Added Rectangle API \([rect.ts](src/math/rect.ts)\)
+- Added Vector 2 API \([vec2.ts](src/math/vec2.ts)\)
 - Added Showcase Landing Page available [here](https://twisterjs.com/).
 
 ## [v0.1.0] - 2025-12-23